Task 0

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,20 +9,19 @@
  */
 size_t print_list(const list_t *h)
 {
-	int count = 0;
-
-	while (h)
-	{
-		if (h->str == NULL)
-		{
-			printf("[0] (nil)\n");
-		}
-		else
-		{
-		printf("[%u] %s\n", h->len, h->str);
-		}
-		count++;
-		h = h->next;
-	}
-	return (count);
+size_t count = 0;
+while (h)
+{
+if (h->str == NULL)
+{
+printf("[0] (nil)\n");
+}
+else
+{
+printf("[%u] %s\n", h->len, h->str);
+}
+count++;
+h = h->next;
+}
+return (count);
 }
